+++ packages/components/src/components/dropdown-select/dropdown-select.tsx
@@ -248,9 +248,8 @@ export class DropdownSelect {
   }
 
   bringIntoView(index) {
-    const options: NodeListOf<HTMLElement> = this.listboxEl.querySelectorAll(
-      '[role=option]'
-    );
+    const options: NodeListOf<HTMLElement> =
+      this.listboxEl.querySelectorAll('[role=option]');
 
     if (hasOverflow(this.listboxEl)) {
       keepInView(options[index], this.listboxEl);
@@ -465,17 +464,11 @@ export class DropdownSelect {
             </div>
           </div>
 
-          {!!this.helperText && (
-            <div
-              part="meta"
-              id={helperTextId}
-              aria-live="polite"
-              aria-relevant="additions removals"
-            >
-              {!!this.helperText && (
-                <div part="helper-text">{this.helperText}</div>
-              )}
-            </div>
+          {this.helperText && (
+            <scale-helper-text
+              helperText={this.helperText}
+              variant={this.invalid ? 'danger' : 'neutral'}
+            ></scale-helper-text>
           )}
         </div>
       </Host>
@@ -493,7 +486,8 @@ export class DropdownSelect {
       this.transparent && 'transparent',
       this.invalid && `invalid`,
       this.currentIndex > -1 && `steal-focus`,
-      animated && 'animated'
+      animated && 'animated',
+      this.helperText && 'has-helper-text'
     );
   }
 }