+++ packages/components/src/components/notification-toast/notification-toast.tsx
@@ -249,8 +249,8 @@ export class NotificationToast {
   };
 
   getToastHeightWithOffset() {
-    const toastHeight = this.element.shadowRoot.querySelector('.toast')
-      .scrollHeight;
+    const toastHeight =
+      this.element.shadowRoot.querySelector('.toast').scrollHeight;
     this.toastHeightWithOffset = toastHeight + this.positionVertical;
   }