+++ packages/components/src/components/text-field/text-field.tsx
@@ -54,14 +54,17 @@ export class TextField {
   @Prop() name?: string = '';
   /** Input label */
   @Prop() label: string = '';
-  /** (optional) Input size */
-  @Prop() size?: string = '';
+  /** @deprecated - css overwrite should replace size */
+  @Prop() size?: string;
   /** (optional) Input helper text */
   @Prop() helperText?: string = '';
   /** @deprecated - invalid should replace status */
   @Prop() status?: string = '';
   /** (optional) Input status */
   @Prop() invalid?: boolean = false;
+  /** (optional) Variant */
+  @Prop() variant?: 'informational' | 'warning' | 'danger' | 'success' =
+    'informational';
   /** (optional) Input text string max length */
   @Prop() maxLength?: number;
   /** (optional) Input text string min length */
@@ -148,6 +151,15 @@ export class TextField {
         tag: 'deprecated',
         message:
           'Property "status" is deprecated. Please use the "invalid" property!',
+        type: 'warn',
+        source: this.hostElement,
+      });
+    }
+    if (this.size) {
+      statusNote({
+        tag: 'deprecated',
+        message:
+          'Property "size" is deprecated. Please use css overwrites for a small version!',
         type: 'warn',
         source: this.hostElement,
       });
@@ -247,7 +259,6 @@ export class TextField {
             {...(this.helperText ? ariaDescribedByAttr : {})}
             {...(numericTypes.includes(this.type) ? { step: this.step } : {})}
           />
-
           {(!!this.helperText || !!this.counter) && (
             <div
               class="text-field__meta"
@@ -255,9 +266,6 @@ export class TextField {
               aria-live="polite"
               aria-relevant="additions removals"
             >
-              {!!this.helperText && (
-                <div class="text-field__helper-text">{this.helperText}</div>
-              )}
               {this.counter && (
                 <div class="text-field__counter">
                   {!!this.value ? String(this.value).length : 0} /{' '}
@@ -265,6 +273,12 @@ export class TextField {
                 </div>
               )}
             </div>
+          )}
+          {this.helperText && (
+            <scale-helper-text
+              helperText={this.helperText}
+              variant={this.invalid ? 'danger' : this.variant}
+            ></scale-helper-text>
           )}
         </div>
       </Host>
@@ -285,8 +299,9 @@ export class TextField {
       this.disabled && `text-field--disabled`,
       this.transparent && 'text-field--transparent',
       this.status && `text-field--status-${this.status}`,
-      this.invalid && `text-field--status-error`,
-      this.size && `text-field--size-${this.size}`,
+      this.invalid && `text-field--variant-danger`,
+      this.variant && `text-field--variant-${this.variant}`,
+      this.helperText && `text-field--helper-text`,
       this.readonly && `text-field--readonly`,
       animated && 'animated'
     );