+++ packages/visual-tests/src/date-picker.visual.spec.js
@@ -33,7 +33,6 @@ describe('DatePicker', () => {
       ['helper-text'],
       ['with-error'],
       ['disabled'],
-      ['small'],
       ['date-range-picker'],
     ])('%p', async (variant) => {
       await global.page.goto(