+++ packages/visual-tests/src/switch.visual.spec.js
@@ -1,3 +1,4 @@
+// FIXME
 describe('Switch', () => {
   describe.each(['light', 'dark'])('%p', (mode) => {
     beforeAll(async () => {
@@ -8,30 +9,36 @@ describe('Switch', () => {
       ['standard-disabled'],
       ['selected'],
       ['selected-disabled'],
-    ])('%p', async (variant) => {
+    ])('story %p', async (variant) => {
       await global.runSetup(`components-switch--${variant}`);
       await global.visualCheck();
     });
     // hover, active
-    test.each([['standard'], ['selected']])('%p', async (variant) => {
+    test.each([['standard'], ['selected']])('hover %p', async (variant) => {
       await global.runSetup(`components-switch--${variant}`);
       const firstButton = await global.page.evaluateHandle(
         'document.querySelector("#root > scale-switch > div")'
       );
       await firstButton.hover();
       await global.visualCheck();
+    });
+    test.each([['standard'], ['selected']])('active %p', async (variant) => {
+      await global.runSetup(`components-switch--${variant}`);
+      const firstButton = await global.page.evaluateHandle(
+        'document.querySelector("#root > scale-switch > div")'
+      );
       await global.page.mouse.move(25, 25);
       await global.page.mouse.down();
       await global.visualCheck();
     });
     // focus
-    test.each([['standard'], ['selected']])('%p', async (variant) => {
+    test.each([['standard'], ['selected']])('focus %p', async (variant) => {
       await global.runSetup(`components-switch--${variant}`);
       await global.page.keyboard.press('Tab');
       await global.visualCheck();
     });
     // click
-    test.each([['standard'], ['selected']])('%p', async (variant) => {
+    test.each([['standard'], ['selected']])('click %p', async (variant) => {
       await global.runSetup(`components-switch--${variant}`);
       const firstButton = await global.page.evaluateHandle(
         'document.querySelector("#root > scale-switch > div .switch__wrapper")'