+++ packages/visual-tests/src/tab-navigation.visual.spec.js
@@ -5,13 +5,29 @@ describe('TabNavigation', () => {
     });
     test.each([
       ['text-icon'],
-      ['text-icon-small'],
       ['text-only'],
-      ['text-only-small'],
       ['disabled-tabs'],
+      ['large-text-only'],
+      ['large-text-icon'],
     ])('%p', async (variant) => {
-      await global.runSetup(`components-tab-navigation--${variant}`);
-      await global.visualCheck();
+      await page.goto(
+        `http://host.docker.internal:3123/iframe.html?id=components-tab-navigation--${variant}&viewMode=story`
+      );
+      await page.waitForSelector('html.hydrated');
+
+      const previewHtml = await page.$('body');
+      await page.evaluate(() => {
+        [
+          '--telekom-motion-duration-immediate',
+          '--telekom-motion-duration-transition',
+          '--telekom-motion-duration-animation',
+          '--telekom-motion-duration-animation-deliberate',
+        ].forEach((transitionSpeed) => {
+          document.body.style.setProperty(transitionSpeed, '0s');
+        });
+      });
+
+      expect(await previewHtml.screenshot()).toMatchImageSnapshot();
     });
     // hover, focus, active
     test.each([['text-icon'], ['text-only']])('%p', async (variant) => {